finished refac of api calls being put into apiHelper folders
